Fix blank `Job.public_slug`
`Job.public_slug` is supposed to be able to be blank, but the uniqueness
constraint means that there can be only one blank ('')  slug.

Fix this by changing "" -> None in `clean()`, and return None
for the `public_url` if there is no slug or if there is no valid reverse
url mapping.