Fix public job view content links
This fixes the content links in the public job view (based on the public slug).

Previously, the links in the content portions of the newsletter are changed to link hashes, but the public view didn’t have a mail_hash portion, so the links resembled `link//<link_hash>/` which is an invalid url.

Now, there’s a new `link/public/<link_hash>/` url which can create a proxy Mail object for the purposes of rendering the url to redirect to.